feat: move K8sWatcher to __init__.py
